requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
dateparser==1.0.0
gevent==21.8.0
greenlet==1.1.1
idna==3.2
install==1.3.4
multidict==5.1.0
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.21
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websocket==0.2.1
websockets==9.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0

python-dotenv~=0.19.0
Telethon~=1.23.0